requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
backcall==0.2.0
cycler==0.10.0
decorator==5.0.5
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.13
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.1
nest-asyncio==1.5.1
numpy==1.20.2
pandas==1.2.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
tkinter
PyArrow >= 4.0.0
